chore: move `CacheContainer` to container.rs
